requirements.txt
langchain_anthropic==0.1.13
openai==1.30.3
plotly==5.15.0
pyperclip==1.8.2
python-dotenv==1.0.1
python_dateutil==2.8.2
Requests==2.32.2
streamlit==1.33.0
streamlit_extras==0.4.2
streamlit_keyup==0.2.4
streamlit_server_state==0.17.1
streamlit_supabase_auth==1.0.2
supabase==2.4.6
tiktoken==0.6.0
cryptography
google-generativeai
groq
transformers
textstat
sentence_transformers
scipy
